requirements.txt
langchain_community==0.3.4
tiktoken==0.8.0
langchain-openai==0.2.4
langchainhub==0.1.21
chromadb==0.5.16
langchain==0.3.6
python-dotenv==1.0.0
beautifulsoup4==4.12.3
unstructured==0.16.3
Markdown==3.7
fastapi==0.115.4
uvicorn==0.22.0